backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
opencv-python==4.7.0.72
pytesseract==0.3.9
numpy==1.24.2
requests==2.28.2
python-dotenv==1.0.0